Move test factories into tests/common/
In preparation for putting more generic support modules in
tests/common/, this commit moves the test factories module.

I've also removed the factories_test.py file. This is for a couple of
reasons:

- we don't test test code (or if we did, would we test test test code?).
- factories are tested in the context where they are used (i.e. in
  regular tests).
- having factories_test.py in `h/` while `factories.py` lives at the top
  level or in `common/` doesn't make a lot of sense -- and I'm not
  inclined to start adding tests to `tests/common/` across the board.
- not explicitly testing `factories.py` means it is more obvious from
  test coverage if parts of it are not being used and can be removed.
