+++ tests/h/feeds/rss_test.py
@@ -5,7 +5,7 @@ import mock
 from h import models
 from h.feeds import rss
 
-from ... import factories
+from ...common import factories
 
 
 def _annotation_url():
